requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
djangorestframework==3.11.0
idna==2.9
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
python-magic==0.4.15
python-resize-image==1.1.19
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
